+++ client/socket/game.js
@@ -1,6 +1,7 @@
 // put any game socket listening in here
 
 export default socket => {
+  // The old brute force game listener
   // whenever the server sends the game state
   // socket.on('send-game-state', gameState => {
   //   // get map
