+++ server/socket/game.js
@@ -1,0 +1,19 @@
+const lobbies = require('../lobbyer')
+
+// actual game stuff
+const game = socket => {
+  //   /**
+  //    * Person places a boat
+  //    */
+  //   socket.on('boat-add', boat => {
+  //     const playerToUpdate = store
+  //       .getState()
+  //       .players.find(p => p.socketId === boat.socketId)
+  //     store.dispatch(
+  //       addBoat(boat.socketId, new Boat(playerToUpdate.color, boat.x, boat.y))
+  //     )
+  //     io.emit('send-game-state', store.getState())
+  //   })
+}
+
+module.exports = game
